introduce module style launching of manager
