+++ src/cocotb_vivado/__init__.py
@@ -1,11 +1,6 @@
-import sys
-import importlib
-import traceback
-import os
+#!/usr/bin/env python3
 
-sys.modules["cocotb.simulator"] = importlib.import_module("cocotb_vivado.stub.simulator")
-
-import cocotb
+from .__main__ import run
 
 # monkey-patch the clock & trigger layer
 import cocotb.clock
@@ -16,22 +11,3 @@ cocotb.triggers.RisingEdge = clock_scheduler.RisingEdge
 cocotb.triggers.FallingEdge = clock_scheduler.FallingEdge
 cocotb.triggers.Edge = clock_scheduler.Edge
 
-from .stub.mgr import Mgr
-
-
-def run(module, xsim_design, top_level_lang):
-    if top_level_lang != "verilog":
-        raise Exception("Only verilog supported as top level languge")
-
-    os.environ["MODULE"] = module
-
-    mgr = Mgr.init(xsim_design)
-
-    cocotb._initialise_testbench([])
-
-    mgr.run()
-
-    mgr.close()
-
-    if cocotb.regression_manager.failures:
-        exit(1)
